+++ src/commands/test-web-module.js
@@ -4,9 +4,5 @@ import karmaServer from '../karmaServer'
 import type {ErrBack} from '../types'
 
 export default function testWebModule(args: Object, cb: ErrBack) {
-  karmaServer(args, {
-    babel: {
-      stage: 1
-    }
-  }, cb)
+  karmaServer(args, {}, cb)
 }