+++ src/inferno/inferno-preset.js
@@ -1,6 +1,8 @@
-module.exports = {
-  plugins: [
-    require.resolve('babel-plugin-syntax-jsx'),
-    require.resolve('babel-plugin-inferno'),
-  ]
+module.exports = function() {
+  return {
+    plugins: [
+      require.resolve('babel-plugin-syntax-jsx'),
+      require.resolve('babel-plugin-inferno'),
+    ]
+  }
 }