+++ src/preact/preact-preset.js
@@ -1,5 +1,7 @@
-module.exports = {
-  plugins: [
-    [require.resolve('babel-plugin-transform-react-jsx'), {pragma: 'h'}]
-  ]
+module.exports = function() {
+  return {
+    plugins: [
+      [require.resolve('@babel/plugin-transform-react-jsx'), {pragma: 'h'}]
+    ]
+  }
 }