+++ src/react/react-dev-preset.js
@@ -1,6 +1,0 @@
-module.exports = {
-  plugins: [
-    require.resolve('babel-plugin-transform-react-jsx-source'),
-    require.resolve('babel-plugin-transform-react-jsx-self'),
-  ]
-}