+++ tests/commands/projects-test.js
@@ -207,15 +207,15 @@ describe('sample projects', function() {
 
     it('ES5 build transpiles to a cherry-picked version', () => {
       expect(es5)
-        .toInclude("require('react-bootstrap/lib/Col')")
-        .toInclude("require('react-bootstrap/lib/Grid')")
-        .toInclude("require('react-bootstrap/lib/Row')")
+        .toInclude('require("react-bootstrap/lib/Col")')
+        .toInclude('require("react-bootstrap/lib/Grid")')
+        .toInclude('require("react-bootstrap/lib/Row")')
     })
     it('ES5 build has propType declarations wrapped in an environment check', () => {
       expect(es5).toInclude('CherryPicker.propTypes = process.env.NODE_ENV !== "production" ? {')
     })
     it('ES5 build includes a CommonJS interop export', () => {
-      expect(es5).toInclude("module.exports = exports['default']")
+      expect(es5).toInclude('module.exports = exports.default')
     })
     it('ES5 build ignores co-located test files and directories', () => {
       expect(glob.sync('*', {cwd: path.resolve('lib')})).toEqual([
@@ -224,9 +224,9 @@ describe('sample projects', function() {
     })
     it('ES modules build transpiles to a cherry-picked version', () => {
       expect(esModules)
-        .toInclude("import _Col from 'react-bootstrap/lib/Col'")
-        .toInclude("import _Grid from 'react-bootstrap/lib/Grid'")
-        .toInclude("import _Row from 'react-bootstrap/lib/Row'")
+        .toInclude('import _Col from "react-bootstrap/lib/Col"')
+        .toInclude('import _Grid from "react-bootstrap/lib/Grid"')
+        .toInclude('import _Row from "react-bootstrap/lib/Row"')
     })
     it('ES module build has propType declarations wrapped in an environment check', () => {
       expect(esModules).toInclude('CherryPicker.propTypes = process.env.NODE_ENV !== "production" ? {')
@@ -272,12 +272,12 @@ describe('sample projects', function() {
     it('creates split bundles, vendor bundles, copies public subdirs and includes font resources', () => {
       let files = stripHashes((glob.sync('**/*', {cwd: path.resolve('dist')}))).sort()
       expect(files).toEqual([
-        '1.js',
-        '1.js.map',
-        '2.css',
-        '2.css.map',
-        '2.js',
-        '2.js.map',
+        '3.css',
+        '3.css.map',
+        '3.js',
+        '3.js.map',
+        '4.js',
+        '4.js.map',
         'app.css',
         'app.css.map',
         'app.js',