+++ tests/createKarmaConfig-test.js
@@ -138,8 +138,8 @@ describe('createKarmaConfig()', () => {
   it('includes polyfill and default test files pattern', () => {
     let config = createKarmaConfig({}, {}, {}, {})
     expect(config.files).toEqual([
-      require.resolve('babel-polyfill/dist/polyfill.js'),
-      '+(src|test?(s))/**/*+(-test|.spec|.test).js',
+      require.resolve('@babel/polyfill/dist/polyfill.js'),
+      '+(src|test?(s))/**/*+(.spec|.test).js',
     ])
   })
   it('includes default mocha reporter config', () => {